moved decompositions to separate file
