+++ tests/core/methods/test_bug.py
@@ -16,20 +16,17 @@ import numpy as np
 from scipy.linalg import expm
 
 from mqt.yaqs.core.data_structures.networks import MPO, MPS
-from mqt.yaqs.core.data_structures.simulation_parameters import Observable, PhysicsSimParams
+from mqt.yaqs.core.data_structures.simulation_parameters import PhysicsSimParams
 from mqt.yaqs.core.methods.bug import (
-    _build_basis_change_tensor,  # noqa: PLC2701
-    _choose_stack_tensor,  # noqa: PLC2701
-    _find_new_q,  # noqa: PLC2701
-    _left_qr,  # noqa: PLC2701
-    _local_update,  # noqa: PLC2701
-    _prepare_canonical_site_tensors,  # noqa: PLC2701
-    _right_qr,  # noqa: PLC2701
-    _right_svd,  # noqa: PLC2701
-    _truncated_right_svd,  # noqa: PLC2701
     bug,
+    build_basis_change_tensor,
+    choose_stack_tensor,
+    find_new_q,
+    local_update,
+    prepare_canonical_site_tensors,
     truncate,
 )
+from mqt.yaqs.core.methods.decompositions import right_qr
 from mqt.yaqs.core.methods.tdvp import update_left_environment
 
 if TYPE_CHECKING:
@@ -90,56 +87,6 @@ def random_mpo(shapes: list[tuple[int, int, int, int]]) -> MPO:
     return mpo
 
 
-def test_right_qr() -> None:
-    """Tests the right qr decomposition.
-
-    Ensures that it produces tensors of the correct shape and a unitary tensor.
-    Also checks that the decomposition is actually the original tensor.
-    """
-    shape = (2, 3, 4)
-    tensor = crandn(shape)
-    q_tensor, r_matrix = _right_qr(tensor)
-    assert q_tensor.ndim == 3
-    assert r_matrix.ndim == 2
-    assert q_tensor.shape[0] == shape[0]
-    assert q_tensor.shape[1] == shape[1]
-    assert r_matrix.shape[1] == shape[2]
-    assert q_tensor.shape[2] == r_matrix.shape[0]
-    # Check that q_tensor is unitary
-    iden = np.eye(q_tensor.shape[2])
-    q_matrix = q_tensor.reshape(q_tensor.shape[0] * q_tensor.shape[1], -1)
-    assert np.allclose(q_matrix.conj().T @ q_matrix, iden)
-    # Check that qr = tensor
-    contr = np.tensordot(q_tensor, r_matrix, axes=(2, 0))
-    assert np.allclose(contr, tensor)
-
-
-def test_left_qr() -> None:
-    """Tests the left qr decomposition.
-
-    Ensures that it produces tensors of the correct shape and a unitary tensor.
-    Also checks that the decomposition is actually the original tensor.
-    """
-    shape = (2, 3, 4)
-    tensor = crandn(shape)
-    q_tensor, r_matrix = _left_qr(tensor)
-    assert q_tensor.ndim == 3
-    assert r_matrix.ndim == 2
-    assert q_tensor.shape[0] == shape[0]
-    assert q_tensor.shape[2] == shape[2]
-    assert r_matrix.shape[0] == shape[1]
-    assert q_tensor.shape[1] == r_matrix.shape[1]
-    # Check that q_tensor is unitary
-    iden = np.eye(q_tensor.shape[1])
-    q_matrix = q_tensor.transpose(0, 2, 1)
-    q_matrix = q_matrix.reshape(-1, q_tensor.shape[1])
-    assert np.allclose(q_matrix.T.conj() @ q_matrix, iden)
-    # Check that qr = tensor
-    contr = np.tensordot(q_tensor, r_matrix, axes=(1, 1))
-    contr = contr.transpose(0, 2, 1)
-    assert np.allclose(contr, tensor)
-
-
 def test_prepare_canonical_site_tensors_single_site() -> None:
     """Tests the preparation for a single site MPS.
 
@@ -152,7 +99,7 @@ def test_prepare_canonical_site_tensors_single_site() -> None:
     mpo_tensor = crandn(2, 2, 1, 1)
     mpo = MPO()
     mpo.init_custom([mpo_tensor])
-    canon_sites, left_envs = _prepare_canonical_site_tensors(mps, mpo)
+    canon_sites, left_envs = prepare_canonical_site_tensors(mps, mpo)
     assert mps.almost_equal(ref_mps)
     assert len(left_envs) == 1
     assert len(canon_sites) == 1
@@ -176,7 +123,7 @@ def test_prepare_canonical_site_tensors_three_sites() -> None:
     mpo_tensors = [crandn(shape) for shape in shapes2]
     mpo = MPO()
     mpo.init_custom(mpo_tensors, transpose=False)
-    canon_sites, left_envs = _prepare_canonical_site_tensors(mps, mpo)
+    canon_sites, left_envs = prepare_canonical_site_tensors(mps, mpo)
     assert mps.almost_equal(ref_mps)
     assert len(left_envs) == 3
     assert len(canon_sites) == 3
@@ -187,13 +134,13 @@ def test_prepare_canonical_site_tensors_three_sites() -> None:
     assert np.allclose(correct_env, left_envs[0])
     assert np.allclose(correct_canon, canon_sites[0])
     # Site 1
-    q_last, r_matrix = _right_qr(mps_tensors[0])
+    q_last, r_matrix = right_qr(mps_tensors[0])
     correct_canon = np.tensordot(r_matrix, mps_tensors[1], axes=(1, 1)).transpose(1, 0, 2)
     correct_env = update_left_environment(q_last, q_last, mpo_tensors[0], left_envs[0])
     assert np.allclose(correct_env, left_envs[1])
     assert np.allclose(correct_canon, canon_sites[1])
     # Site 2
-    q_last, r_matrix = _right_qr(correct_canon)
+    q_last, r_matrix = right_qr(correct_canon)
     correct_canon = np.tensordot(r_matrix, mps_tensors[2], axes=(1, 1)).transpose(1, 0, 2)
     correct_env = update_left_environment(q_last, q_last, mpo_tensors[1], left_envs[1])
     assert np.allclose(correct_env, left_envs[2])
@@ -211,7 +158,7 @@ def test_choose_stack_tensor_last_site() -> None:
     mps = MPS(num_sites, tensors=mps_tensors)
     canon_center_tensors = [crandn(2, 3, 4) for _ in range(num_sites)]
     # Found tensor
-    found_tensor = _choose_stack_tensor(num_sites - 1, canon_center_tensors, mps)
+    found_tensor = choose_stack_tensor(num_sites - 1, canon_center_tensors, mps)
     assert np.allclose(mps_tensors[-1], found_tensor)
 
 
@@ -226,7 +173,7 @@ def test_choose_stack_tensor_middle_site() -> None:
     mps = MPS(num_sites, tensors=mps_tensors)
     canon_center_tensors = [crandn(2, 3, 4) for _ in range(num_sites)]
     # Found tensor
-    found_tensor = _choose_stack_tensor(1, canon_center_tensors, mps)
+    found_tensor = choose_stack_tensor(1, canon_center_tensors, mps)
     assert np.allclose(canon_center_tensors[1], found_tensor)
 
 
@@ -238,7 +185,7 @@ def test_find_new_q() -> None:
     """
     old_tensor = crandn(2, 3, 5)
     new_tensor = crandn(2, 4, 5)
-    q_tensor = _find_new_q(old_tensor, new_tensor)
+    q_tensor = find_new_q(old_tensor, new_tensor)
     # Test shape
     assert q_tensor.ndim == 3
     assert q_tensor.shape[0] == 2
@@ -259,7 +206,7 @@ def test_build_basis_change_tensor() -> None:
     old_q = crandn(2, 3, 4)
     new_q = crandn(2, 7, 5)
     old_m = crandn(4, 5)
-    basis_change = _build_basis_change_tensor(old_q, new_q, old_m)
+    basis_change = build_basis_change_tensor(old_q, new_q, old_m)
     assert basis_change.ndim == 2
     assert basis_change.shape[0] == 3
     assert basis_change.shape[1] == 7
@@ -280,14 +227,14 @@ def test_local_update() -> None:
     mps.set_canonical_form(0)
     ref_mps = deepcopy(mps)
     mpo = random_mpo([(2, 2, 1, 3), (2, 2, 3, 4), (2, 2, 4, 1)])
-    canon_sites, left_envs = _prepare_canonical_site_tensors(mps, mpo)
+    canon_sites, left_envs = prepare_canonical_site_tensors(mps, mpo)
     ref_canon_sites = deepcopy(canon_sites)
     right_block = np.eye(5).reshape(5, 1, 5)
     site = 2
     right_m_block = np.eye(5)
     sim_params = PhysicsSimParams(observables=[], elapsed_time=1)
     # Perform the local update
-    result = _local_update(
+    result = local_update(
         mps, mpo, left_envs, right_block, canon_sites, site, right_m_block, sim_params, numiter_lanczos=25
     )
     # General Change Check
@@ -301,92 +248,6 @@ def test_local_update() -> None:
     assert len(result) == 2
     assert result[0].shape == (3, 6)
     assert result[1].shape == (6, 4, 6)
-
-
-def test_right_svd() -> None:
-    """Test that the svd produces the correct shapes and tensors."""
-    tensor = crandn(2, 3, 4)
-    u_tensor, s_vec, v_matrix = _right_svd(tensor)
-    # Check shapes
-    assert u_tensor.shape[0] == 2
-    assert u_tensor.shape[1] == 3
-    assert v_matrix.shape[1] == 4
-    assert u_tensor.shape[2] == s_vec.shape[0]
-    assert s_vec.shape[0] == v_matrix.shape[0]
-    # Check that u_tensor is unitary
-    iden = np.eye(u_tensor.shape[2])
-    result = np.tensordot(u_tensor, u_tensor.conj(), axes=([0, 1], [0, 1]))
-    assert np.allclose(result, iden)
-    # Check that v_matrix is unitary
-    iden = np.eye(v_matrix.shape[1])
-    result = v_matrix.conj().T @ v_matrix
-    assert np.allclose(result, iden)
-    # Check that svd = tensor
-    contr = np.tensordot(u_tensor, np.diag(s_vec) @ v_matrix, axes=(2, 0))
-    assert np.allclose(contr, tensor)
-
-
-def test_truncated_right_svd_thresh() -> None:
-    """Test that the tensor is correctly truncated."""
-    # Placeholder
-    measurements = [Observable("z", site) for site in range(0)]
-    sim_params = PhysicsSimParams(
-        measurements,
-        elapsed_time=0.2,
-        dt=0.1,
-        sample_timesteps=True,
-        num_traj=1,
-        max_bond_dim=4,
-        threshold=0.2,
-        order=1,
-    )
-    s_vector_i = np.array([1, 0.5, 0.1, 0.01])
-    u_tensor_i, _ = _right_qr(crandn(2, 3, 4))
-    v_matrix_i, _ = np.linalg.qr(crandn(4, 4))
-    tensor = np.tensordot(u_tensor_i, np.diag(s_vector_i) @ v_matrix_i, axes=(2, 0))
-
-    # Thus the values 0.1 and 0.01 should be truncated
-    u_tensor, s_vector, v_matrix = _truncated_right_svd(tensor, sim_params)
-    # Check shapes
-    assert u_tensor.shape[0] == 2
-    assert u_tensor.shape[1] == 3
-    assert v_matrix.shape[1] == 4
-    assert u_tensor.shape[2] == 2
-    assert v_matrix.shape[0] == 2
-    assert s_vector.shape[0] == 2
-    assert np.allclose(s_vector, s_vector_i[:2])
-
-
-def test_truncated_right_svd_maxbd() -> None:
-    """Test that the tensor is correctly truncated."""
-    # Placeholder
-    measurements = [Observable("z", site) for site in range(0)]
-    sim_params = PhysicsSimParams(
-        measurements,
-        elapsed_time=0.2,
-        dt=0.1,
-        sample_timesteps=True,
-        num_traj=1,
-        max_bond_dim=3,
-        threshold=1e-4,
-        order=1,
-    )
-
-    s_vector_i = np.array([1, 0.5, 0.1, 0.01])
-    u_tensor_i, _ = _right_qr(crandn(2, 3, 4))
-    v_matrix_i, _ = np.linalg.qr(crandn(4, 4))
-    tensor = np.tensordot(u_tensor_i, np.diag(s_vector_i) @ v_matrix_i, axes=(2, 0))
-
-    # Thus the value 0.01 should be truncated
-    u_tensor, s_vector, v_matrix = _truncated_right_svd(tensor, sim_params)
-    # Check shapes
-    assert u_tensor.shape[0] == 2
-    assert u_tensor.shape[1] == 3
-    assert v_matrix.shape[1] == 4
-    assert u_tensor.shape[2] == sim_params.max_bond_dim
-    assert sim_params.max_bond_dim == v_matrix.shape[0]
-    assert sim_params.max_bond_dim == s_vector.shape[0]
-    assert np.allclose(s_vector, s_vector_i[: sim_params.max_bond_dim])
 
 
 def test_truncate_no_truncation() -> None:
